Ajustes

+++ src/main/kotlin/br/com/unip/carrinho/dto/ItemDTO.kt
@@ -1,3 +1,3 @@
 package br.com.unip.carrinho.dto
 
-class ItemDTO(val id: String, val nome: String, val observacoes: String?, val quantidade: Long, val valor: Double)
+class ItemDTO(val id: String, val nome: String, val observacoes: String?, val quantidade: Int, val valor: Double)
