+++ src/main/kotlin/br/com/unip/carrinho/repository/entity/Item.kt
@@ -1,3 +1,3 @@
 package br.com.unip.carrinho.repository.entity
 
-class Item(val id: String, val produto: String, val observacoes: String?, val quantidade: Long, val valor: Double)
+class Item(val id: String, val produto: String, val observacoes: String?, val quantidade: Int, val valor: Double)
