+++ src/main/kotlin/br/com/unip/carrinho/repository/entity/ProdutoCarrinho.kt
@@ -6,11 +6,11 @@ class ProdutoCarrinho {
 
     var observacoes: String? = ""
 
-    var quantidade: Long = 0
+    var quantidade: Int = 0
 
     constructor()
 
-    constructor(produto: Produto, observacoes: String?, quantidade: Long) {
+    constructor(produto: Produto, observacoes: String?, quantidade: Int) {
         this.produto = produto
         this.observacoes = observacoes
         this.quantidade = quantidade
